// packages/apps/cocm-registry/src/data/schemas/FinancialModel.ts
type FinancialModelDBRecordTypeV1 = {
    schema_version: string
    id: string
    user_id: string
    name: string
    description: string
    address_state: string
    full_time_pcps: number
    pct_pcps_participation: number
    monthly_pcp_referrals: number
    pct_patients_accepting_cocm: number
    pct_patients_graduating: number
    payer_distribution_medicare: number
    payer_distribution_medicare_cocm_expected: number
    payer_distribution_medicaid: number
    payer_distribution_medicaid_cocm_expected: number
    payer_distribution_commercial_cocm: number
    payer_distribution_commercial_cocm_cocm_expected: number
    payer_distribution_commercial_no_cocm: number
    caseload_complexity_pct_patients_suicidality: number
    caseload_complexity_pct_patients_prior_ed: number
    caseload_complexity_pct_patients_substance_abuse: number
    caseload_complexity_pct_patients_above_moderate: number
    care_manager_salary: number
    care_manager_benefits: number
    psychiatrist_salary: number
    psychiatrist_benefits: number
    service_unit_direct_bhm__warm_connection_over_16: number
    service_unit_direct_bhm__initial_assessment_visit: number
    service_unit_direct_bhm__follow_up_visit: number
    service_unit_direct_bhm__group_treatment: number
    service_unit_other_bhm__warm_connection_under1_16: number
    service_unit_other_bhm__outreach_attempts: number
    service_unit_other_bhm__telephone_visit: number
    service_unit_other_bhm__caseload_patient_and_psych_consult: number
    service_unit_other_bhm__team_communication: number
    service_unit_other_bhm__registry_management: number
    service_unit_admin_bhm__charting: number
    service_unit_admin_bhm__other: number
    service_unit_direct_cp_treatment_visit: number
    service_unit_direct_cp_follow_up: number
    service_unit_other_cp_registry_review: number
    service_unit_other_cp_direct_pcp_communication: number
    service_unit_other_cp_caseload_review: number
    service_unit_other_cp_charting: number
    service_unit_other_cp_other: number
    billing_rate_not_seen__patients: number
    billing_rate_not_seen__avg_amount: number
    billing_rate_G2214__patients: number
    billing_rate_G2214__avg_amount: number
    billing_rate_99492__patients: number
    billing_rate_99492__avg_amount: number
    billing_rate_99492_99494__patients: number
    billing_rate_99492_99494__avg_amount: number
    billing_rate_99492_99494x2__patients: number
    billing_rate_99492_99494x2__avg_amount: number
    billing_rate_99493__patients: number
    billing_rate_99493__avg_amount: number
    billing_rate_99493_99494__patients: number
    billing_rate_99493_99494__avg_amount: number
    billing_rate_99493_99494x2__patients: number
    billing_rate_99493_99494x2__avg_amount: number
    modified_date: number
    created_date: number
}

export default FinancialModelDBRecordTypeV1